requirements.txt
crewai
langchain_openai
python-dotenv
requests
duckduckgo-search
langchain_community
PyYAML
dashscope
